Add multiagent section
