+++ src/MultiAgentReinforcementLearning/Project.toml
@@ -1,0 +1,22 @@
+name = "MultiAgentReinforcementLearning"
+uuid = "6107ce81-f7a4-4dec-a483-c13c594fca14"
+version = "0.0.1"
+
+[deps]
+ReinforcementLearningBase = "e575027e-6cd6-5018-9292-cdc6200d2b44"
+ReinforcementLearningCore = "de1b191a-4ae0-4afa-a27b-92d07f46b2d6"
+ReinforcementLearningTrajectories = "6486599b-a3cd-4e92-a99a-2cea90cc8c3c"
+
+[compat]
+ReinforcementLearningBase = "0.11"
+ReinforcementLearningCore = "0.9"
+ReinforcementLearningTrajectories = "^0.1.9"
+julia = "1.6"
+
+[extras]
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+ReinforcementLearningEnvironments = "25e41dd2-4622-11e9-1641-f1adca772921"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test", "Random", "ReinforcementLearningEnvironments"]
