+++ src/MultiAgentReinforcementLearning/multiagent_demo.jl
@@ -12,277 +12,19 @@ begin
 	using Random: Random, AbstractRNG
 	# import ReinforcementLearningEnvironments: TicTacToeEnv
 
-	mutable struct TicTacToeEnv1 <: AbstractEnv
-		board::BitArray{3}
-		player::Symbol
-	end
+	
+	
 
-	function (p::RandomPolicy{Nothing,RNG})(env::TicTacToeEnv1) where {RNG<:AbstractRNG}
-		legal_action_space_ = RLBase.legal_action_space(env)
-		if length(legal_action_space_) >= 0
-    		return rand(p.rng, legal_action_space_)
-		else
-			return nothing
-		end
-	end
-
-	
-	import Functors: functor
-	using Setfield: @set
-
-
-	struct MultiAgentPolicy <: AbstractPolicy
-	    agents::NamedTuple
-	
-		function MultiAgentPolicy(agents::NamedTuple)
-			new(agents)
-		end
-	end
-
-	function _run(multiagent_policy::MultiAgentPolicy, env::AbstractEnv, stop_condition, hook, reset_condition)
-
-	    hook(PreExperimentStage(), multiagent_policy, env)
-	    multiagent_policy(PreExperimentStage(), env)
-	    is_stop = false
-	    while !is_stop
-	        reset!(env)
-	        multiagent_policy(PreEpisodeStage(), env)
-	        hook(PreEpisodeStage(), multiagent_policy, env)
-	
-	        while !reset_condition(multiagent_policy, env) # one episode
-				for player in current_player_iterator(env)
-					policy = multiagent_policy[player] # Select appropriate policy
-
-		            policy(PreActStage(), env)
-		            hook(PreActStage(), policy, env)
-
-					action = policy(env)
-		            env(action)
-					
-		            optimise!(policy)
-		
-		            policy(PostActStage(), env)
-		            hook(PostActStage(), policy, env)
-		
-		            if stop_condition(policy, env)
-		                is_stop = true
-		                policy(PreActStage(), env)
-		                hook(PreActStage(), policy, env)
-		                # policy(env)  # let the policy see the last observation # NOTE: This fails for RandomPolicy because set of legal actions is empty, minor tweak to RandomPolicy code resolves this
-		                break
-		            end
-				end
-	        end # end of an episode
-	
-	        if is_terminated(env)
-	            multiagent_policy(PostEpisodeStage(), env)  # let the policy see the last observation
-	            hook(PostEpisodeStage(), multiagent_policy, env)
-	        end
-	    end
-	    multiagent_policy(PostExperimentStage(), env)
-	    hook(PostExperimentStage(), multiagent_policy, env)
-	    hook
-	end
-	# (policy::NamedPolicy)(env::TicTacToeEnv1) = policy.policy.policy(env)
-
-	struct CurrentPlayerIterator
-		env::AbstractEnv
-	end
-
-	function current_player_iterator(env)
-		return CurrentPlayerIterator(env)
-	end
-
-	Base.iterate(current_player_iterator::CurrentPlayerIterator) = (current_player(current_player_iterator.env), current_player_iterator.env)
-	
-	Base.iterate(current_player_iterator::CurrentPlayerIterator, env) = (current_player(current_player_iterator.env), current_player_iterator.env)
-
-	# current_player()
-	
-	function TicTacToeEnv1()
-	    board = BitArray{3}(undef, 3, 3, 3)
-	    fill!(board, false)
-	    board[:, :, 1] .= true
-	    TicTacToeEnv1(board, :Cross)
-	end
-	
-	function RLBase.reset!(env::TicTacToeEnv1)
-	    fill!(env.board, false)
-	    env.board[:, :, 1] .= true
-	end
-	
-	struct TicTacToeInfo
-	    is_terminated::Bool
-	    winner::Union{Nothing,Symbol}
-	end
-	
-	const TIC_TAC_TOE_STATE_INFO = Dict{
-	    TicTacToeEnv1,
-	    NamedTuple{
-	        (:index, :is_terminated, :winner),
-	        Tuple{Int,Bool,Union{Nothing,Symbol}},
-	    },
-	}()
-	
-	Base.hash(env::TicTacToeEnv1, h::UInt) = hash(env.board, h)
-	Base.isequal(a::TicTacToeEnv1, b::TicTacToeEnv1) = isequal(a.board, b.board)
-	
-	Base.to_index(::TicTacToeEnv1, player) = player == :Cross ? 2 : 3
-
-	Base.getindex(p::MultiAgentPolicy, s::Symbol) = p.agents[s]
-	RLBase.action_space(::TicTacToeEnv1, player) = Base.OneTo(9)
-	
-	RLBase.legal_action_space(env::TicTacToeEnv1, p) = findall(legal_action_space_mask(env))
-	
-	function RLBase.legal_action_space_mask(env::TicTacToeEnv1, p)
-	    if is_win(env, :Cross) || is_win(env, :Nought)
-	        falses(9)
-	    else
-	        vec(env.board[:, :, 1])
-	    end
-	end
-	
-	(env::TicTacToeEnv1)(action::Int) = env(CartesianIndices((3, 3))[action])
-	
-	function (env::TicTacToeEnv1)(action::CartesianIndex{2})
-	    env.board[action, 1] = false
-	    env.board[action, Base.to_index(env, current_player(env))] = true
-	end
-
-	state(env::TicTacToeEnv1) = state(env, Observation{Int}(), 1)
-	state(env::TicTacToeEnv1, ::Observation{BitArray{3}}, p) = env.board
-	state(env::TicTacToeEnv1, ::RLBase.AbstractStateStyle) = state(env::TicTacToeEnv1, Observation{Int}(), 1)
-	state(env::TicTacToeEnv1, p) = state(env, Observation{Int}(), p)
-	state_space(env::TicTacToeEnv1, ::Observation{BitArray{3}}, p) = ArrayProductDomain(fill(false:true, 3, 3, 3))
-	state(env::TicTacToeEnv1, ::Observation{Int}, p) =
-	    get_tic_tac_toe_state_info()[env].index
-	RLBase.state_space(env::TicTacToeEnv1, ::Observation{Int}, p) =
-	    Base.OneTo(length(get_tic_tac_toe_state_info()))
-	RLBase.state_space(env::TicTacToeEnv1, ::Observation{String}, p) = fullspace(String)
-
-	state(env::TicTacToeEnv1, ::Observation{String}) = state(env::TicTacToeEnv1, Observation{String}(), 1)
-	
-	function RLBase.state(env::TicTacToeEnv1, ::Observation{String}, p)
-	    buff = IOBuffer()
-	    for i in 1:3
-	        for j in 1:3
-	            if env.board[i, j, 1]
-	                x = '.'
-	            elseif env.board[i, j, 2]
-	                x = 'x'
-	            else
-	                x = 'o'
-	            end
-	            print(buff, x)
-	        end
-	        print(buff, '\n')
-	    end
-	    String(take!(buff))
-	end
-	
-	RLBase.is_terminated(env::TicTacToeEnv1) = get_tic_tac_toe_state_info()[env].is_terminated
-	
-	function RLBase.reward(env::TicTacToeEnv1, player)
-	    if is_terminated(env)
-	        winner = get_tic_tac_toe_state_info()[env].winner
-	        if isnothing(winner)
-	            0
-	        elseif winner === player
-	            1
-	        else
-	            -1
-	        end
-	    else
-	        0
-	    end
-	end
-	
-	function is_win(env::TicTacToeEnv1, player)
-	    b = env.board
-	    p = Base.to_index(env, player)
-	    @inbounds begin
-	        b[1, 1, p] & b[1, 2, p] & b[1, 3, p] ||
-	            b[2, 1, p] & b[2, 2, p] & b[2, 3, p] ||
-	            b[3, 1, p] & b[3, 2, p] & b[3, 3, p] ||
-	            b[1, 1, p] & b[2, 1, p] & b[3, 1, p] ||
-	            b[1, 2, p] & b[2, 2, p] & b[3, 2, p] ||
-	            b[1, 3, p] & b[2, 3, p] & b[3, 3, p] ||
-	            b[1, 1, p] & b[2, 2, p] & b[3, 3, p] ||
-	            b[1, 3, p] & b[2, 2, p] & b[3, 1, p]
-	    end
-	end
-
-
-	function get_tic_tac_toe_state_info()
-	    if isempty(TIC_TAC_TOE_STATE_INFO)
-	        @info "initializing tictactoe state info cache..."
-	        t = @elapsed begin
-	            n = 1
-	            root = TicTacToeEnv1()
-	            TIC_TAC_TOE_STATE_INFO[root] =
-	                (index=n, is_terminated=false, winner=nothing)
-	            walk(root) do env
-	                if !haskey(TIC_TAC_TOE_STATE_INFO, env)
-	                    n += 1
-	                    has_empty_pos = any(view(env.board, :, :, 1))
-	                    w = if is_win(env, :Cross)
-	                        :Cross
-	                    elseif is_win(env, :Nought)
-	                        :Nought
-	                    else
-	                        nothing
-	                    end
-	                    TIC_TAC_TOE_STATE_INFO[env] = (
-	                        index=n,
-	                        is_terminated=!(has_empty_pos && isnothing(w)),
-	                        winner=w,
-	                    )
-	                end
-	            end
-	        end
-	        @info "finished initializing tictactoe state info cache in $t seconds"
-	    end
-	    TIC_TAC_TOE_STATE_INFO
-	end
-
-	RLBase.current_player(env::TicTacToeEnv1) = env.player
-	
-	RLBase.NumAgentStyle(::TicTacToeEnv1) = MultiAgent(2)
-	RLBase.DynamicStyle(::TicTacToeEnv1) = SEQUENTIAL
-	RLBase.ActionStyle(::TicTacToeEnv1) = FULL_ACTION_SET
-	RLBase.InformationStyle(::TicTacToeEnv1) = PERFECT_INFORMATION
-	RLBase.StateStyle(::TicTacToeEnv1) =
-	    (Observation{String}(), Observation{Int}(), Observation{BitArray{3}}())
-	RLBase.RewardStyle(::TicTacToeEnv1) = TERMINAL_REWARD
-	RLBase.UtilityStyle(::TicTacToeEnv1) = ZERO_SUM
-	RLBase.ChanceStyle(::TicTacToeEnv1) = DETERMINISTIC	
 end
 
 # ╔═╡ 8fae8fc4-e817-4dee-823d-e0b0e6489edb
 begin
-	trajectory_1 = Trajectory(
-		CircularArraySARTTraces(; capacity = 1),
-		BatchSampler(1),
-		InsertSampleRatioController(n_inserted = -1),
-	)
-
-	trajectory_2 = Trajectory(
-		CircularArraySARTTraces(; capacity = 1),
-		BatchSampler(1),
-		InsertSampleRatioController(n_inserted = -1),
-	)
 end
 
 # ╔═╡ c3caf177-de58-47a0-badd-d0e15f43b46f
-multiagent_policy = MultiAgentPolicy((Cross = Agent(RandomPolicy(), trajectory_1), Nought = Agent(RandomPolicy(), trajectory_2)))
 
 # ╔═╡ 6da17a4e-93b8-45a1-a20f-b3410b86b1cc
 begin
-	env = TicTacToeEnv1()
-	stop_condition = StopWhenDone()
-	hook = StepsPerEpisode()
-	run(multiagent_policy, env, stop_condition, hook)
-	is_terminated(env)
 end
 
 # ╔═╡ f20dbc62-9f6c-49a0-bb50-02e11ff8be31
